Add an abstract base class for TransformPal.
This allows us to have implementations with different FFT structures.
